refactor(studio): minor cleanup of FunctionList and exports
