+++ apps/studio/components/interfaces/Database/Functions/FunctionsList/FunctionList.tsx
@@ -1,17 +1,9 @@
 import { PermissionAction } from '@supabase/shared-types/out/constants'
 import { includes, noop, sortBy } from 'lodash'
-import { Check, Copy, Edit, Edit2, FileText, Globe, MoreVertical, Trash, X } from 'lucide-react'
+import { Copy, Edit, Edit2, FileText, MoreVertical, Trash } from 'lucide-react'
 import Link from 'next/link'
 import { useRouter } from 'next/router'
-
-import { SIDEBAR_KEYS } from 'components/layouts/ProjectLayout/LayoutSidebar/LayoutSidebarProvider'
-import { ButtonTooltip } from 'components/ui/ButtonTooltip'
-import { useAsyncCheckPermissions } from 'hooks/misc/useCheckPermissions'
-import { useSelectedProjectQuery } from 'hooks/misc/useSelectedProject'
-import { useAiAssistantStateSnapshot } from 'state/ai-assistant-state'
-import { useSidebarManagerSnapshot } from 'state/sidebar-manager-state'
 import {
-  Badge,
   Button,
   DropdownMenu,
   DropdownMenuContent,
@@ -20,12 +12,17 @@ import {
   DropdownMenuTrigger,
   TableCell,
   TableRow,
-  Tooltip,
-  TooltipContent,
-  TooltipTrigger,
 } from 'ui'
-import type { DatabaseFunction } from 'data/database-functions/database-functions-query'
-import type { FunctionApiAccessMap } from 'data/privileges/function-api-access-query'
+
+import { ApiAccessCell, ApiAccessMenuItem } from './ApiAccess'
+import { SIDEBAR_KEYS } from '@/components/layouts/ProjectLayout/LayoutSidebar/LayoutSidebarProvider'
+import { ButtonTooltip } from '@/components/ui/ButtonTooltip'
+import type { DatabaseFunction } from '@/data/database-functions/database-functions-query'
+import type { FunctionApiAccessMap } from '@/data/privileges/function-api-access-query'
+import { useAsyncCheckPermissions } from '@/hooks/misc/useCheckPermissions'
+import { useSelectedProjectQuery } from '@/hooks/misc/useSelectedProject'
+import { useAiAssistantStateSnapshot } from '@/state/ai-assistant-state'
+import { useSidebarManagerSnapshot } from '@/state/sidebar-manager-state'
 
 interface FunctionListProps {
   schema: string
@@ -41,7 +38,7 @@ interface FunctionListProps {
   onToggleApiAccess?: (fn: DatabaseFunction, enable: boolean) => void
 }
 
-const FunctionList = ({
+export function FunctionList({
   schema,
   filterString,
   isLocked,
@@ -53,7 +50,7 @@ const FunctionList = ({
   functions,
   functionApiAccessMap,
   onToggleApiAccess = noop,
-}: FunctionListProps) => {
+}: FunctionListProps) {
   const router = useRouter()
   const { data: selectedProject } = useSelectedProjectQuery()
   const aiSnap = useAiAssistantStateSnapshot()
@@ -152,9 +149,7 @@ const FunctionList = ({
               </p>
             </TableCell>
             <TableCell className="table-cell">
-              <ApiAccessCell
-                apiAccessData={functionApiAccessMap?.[x.id]}
-              />
+              <ApiAccessCell apiAccessData={functionApiAccessMap?.[x.id]} />
             </TableCell>
             <TableCell className="text-right">
               {!isLocked && (
@@ -259,96 +254,3 @@ const FunctionList = ({
   )
 }
 
-const ApiAccessCell = ({
-  apiAccessData,
-}: {
-  apiAccessData?: FunctionApiAccessMap[number]
-}) => {
-  if (!apiAccessData) {
-    return (
-      <p className="truncate text-foreground-muted">–</p>
-    )
-  }
-
-  if (apiAccessData.apiAccessType === 'none') {
-    return (
-      <Tooltip>
-        <TooltipTrigger>
-          <Badge variant="default">
-            <X size={12} className="mr-1" />
-            No
-          </Badge>
-        </TooltipTrigger>
-        <TooltipContent side="top" className="max-w-xs">
-          Schema is not exposed via the Data API. Enable the schema in API settings to allow API access.
-        </TooltipContent>
-      </Tooltip>
-    )
-  }
-
-  if (apiAccessData.apiAccessType === 'exposed-schema-no-grants') {
-    return (
-      <Tooltip>
-        <TooltipTrigger>
-          <Badge variant="default">
-            <X size={12} className="mr-1" />
-            No
-          </Badge>
-        </TooltipTrigger>
-        <TooltipContent side="top" className="max-w-xs">
-          Function is not accessible via the Data API. Grant EXECUTE privileges to anon or authenticated roles to enable access.
-        </TooltipContent>
-      </Tooltip>
-    )
-  }
-
-  // apiAccessType === 'access'
-  const roles = []
-  if (apiAccessData.privileges.anon) roles.push('anon')
-  if (apiAccessData.privileges.authenticated) roles.push('authenticated')
-
-  return (
-    <Tooltip>
-      <TooltipTrigger>
-        <Badge variant="brand">
-          <Check size={12} className="mr-1" />
-          Yes
-        </Badge>
-      </TooltipTrigger>
-      <TooltipContent side="top" className="max-w-xs">
-        Function is accessible via the Data API for roles: {roles.join(', ')}
-      </TooltipContent>
-    </Tooltip>
-  )
-}
-
-const ApiAccessMenuItem = ({
-  apiAccessData,
-  onToggle,
-}: {
-  apiAccessData?: FunctionApiAccessMap[number]
-  onToggle: (enable: boolean) => void
-}) => {
-  if (!apiAccessData) {
-    return null
-  }
-
-  // If schema is not exposed, don't show the toggle option
-  if (apiAccessData.apiAccessType === 'none') {
-    return null
-  }
-
-  const hasAccess = apiAccessData.apiAccessType === 'access'
-
-  return (
-    <DropdownMenuItem
-      className="space-x-2"
-      onClick={() => onToggle(!hasAccess)}
-    >
-      <Globe size={14} />
-      <p>{hasAccess ? 'Disable API access' : 'Enable API access'}</p>
-    </DropdownMenuItem>
-  )
-}
-
-export default FunctionList
