+++ apps/studio/pages/project/[ref]/database/functions.tsx
@@ -1,13 +1,4 @@
 import { PermissionAction } from '@supabase/shared-types/out/constants'
-
-import FunctionsList from 'components/interfaces/Database/Functions/FunctionsList/FunctionsList'
-import DatabaseLayout from 'components/layouts/DatabaseLayout/DatabaseLayout'
-import DefaultLayout from 'components/layouts/DefaultLayout'
-import { DocsButton } from 'components/ui/DocsButton'
-import NoPermission from 'components/ui/NoPermission'
-import { useAsyncCheckPermissions } from 'hooks/misc/useCheckPermissions'
-import { DOCS_URL } from 'lib/constants'
-import type { NextPageWithLayout } from 'types'
 import { PageContainer } from 'ui-patterns/PageContainer'
 import {
   PageHeader,
@@ -17,6 +8,15 @@ import {
   PageHeaderTitle,
 } from 'ui-patterns/PageHeader'
 import { PageSection, PageSectionContent } from 'ui-patterns/PageSection'
+
+import { FunctionsList } from '@/components/interfaces/Database/Functions/FunctionsList/FunctionsList'
+import DatabaseLayout from '@/components/layouts/DatabaseLayout/DatabaseLayout'
+import DefaultLayout from '@/components/layouts/DefaultLayout'
+import { DocsButton } from '@/components/ui/DocsButton'
+import NoPermission from '@/components/ui/NoPermission'
+import { useAsyncCheckPermissions } from '@/hooks/misc/useCheckPermissions'
+import { DOCS_URL } from '@/lib/constants'
+import type { NextPageWithLayout } from '@/types'
 
 const DatabaseFunctionsPage: NextPageWithLayout = () => {
   const { can: canReadFunctions, isSuccess: isPermissionsLoaded } = useAsyncCheckPermissions(
